requirements.txt
selenium
requests
webdriver_manager
bs4
python-Levenshtein
wordcloud 
matplotlib
pandas
boto3
